pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pap_maker"
dynamic = ["version"]
description = "pip-installable Python PAP file maker."
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
{name = "Giuseppe De Gregorio", email = "degregorio.gius@gmail.com"},
]
keywords = ["papfile", "utility", "packaging"]
classifiers = [
"Programming Language :: Python :: 3",
"License :: OSI Approved :: MIT License",
"Operating System :: OS Independent",
"Development Status :: 3 - Alpha",
]

dependencies = [
"pillow>=12.0.0",
]

[project.scripts]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {file = "src/pap_maker/init.py"}